backend/requirements.txt
flask
flask-cors
spotipy
textblob
requests
numpy
python-dotenv
gunicorn
